# src/main/java/com/restful/product_crud/model/CategoryResponse.java
package com.restful.product_crud.model;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class CategoryResponse {

    private int id;

    private String name;

}